+++ t/request.lisp
@@ -9,13 +9,15 @@
 (plan 1)
 
 
+
 (subtest "refresh-request-p-1"
   (with-session
     (with-request ("/foo/bar")
       (weblocks.session:set-value 'weblocks::last-request-uri
                                   '("foo" "bar"))
       (is (weblocks.request:refresh-request-p)
-          t))))
+          t
+          "Refresh-request-p should return true, because current URI has same tokens as last-request-uri."))))
 
 
 (finalize)